[PLAT-6623] Add filters to prevent spurious kills
Adding filters to avoid killing processes if exe-kill cannot find
the process or if there is no proc entry.  Without the filter,
the script causes problems because the first readlink may be empty
and the second readlink may be empty due to a permission error.

This will cause the script to kill processes for which the exe is
not available to the user.  This kill then proceeds to cause ssh
connections to disconnect.